Bulls and Cows game
Given the secret number ‘secret’ and your friend's guess ‘guess’, return the hint for your
friend's guess.
The hint should be formatted as "xAyB", where x is the number of bulls and y is the
number of cows. Note that both secret and guess may contain duplicate digits.
Constraints:
1 <= secret.length, guess.length <= 1000
secret.length = = guess.length
secret and guess consist of digits only.
Test Case 1:
Input: secret = "1807", guess = "7810"
Output: "1A3B"
Explanation: Bulls are connected with a '|' and cows are underlined:
"1807"
|
"7810"
Test Case 2:
Input: secret = "1123", guess = "0111"
Output: "1A1B"
Explanation: Bulls are connected with a '|' and cows are underlined:
"1123" "1123"
| or |
"0111" "0111"
Note that only one of the two unmatched 1s is counted as a cow since the non-bull digits
can only be rearranged to allow one 1 to be a bull.

Test Case 3:
Input: secret = "9999", guess = "1111"
Output: "0A0B"
Explanation: In this case, there are no bulls (no digits in the correct position) and no cows
(no digits that appear in the secret number but are in the wrong position). This is because
all the digits in the guess ("1111") are not present in the secret number ("9999").